+++ packages/perps-controller/src/constants/index.ts
@@ -9,3 +9,4 @@ export * from './perpsConfig.js';
 export * from './transactionsHistoryConfig.js';
 export * from './performanceMetrics.js';
 export * from './myxConfig.js';
+export * from './lighterConfig.js';